Create hook for navigation params'
